Windows 10 Bloatware / Adware Removal
This is a Powershell Script to Remove Installed Bloatware (Listed) - remember to un-pin any advertisements afterwards such as Minecraft, Adobe Photoshop Elements, Royal Revolt... etc. This PS CMDlet will not un-pin adverts, it only uninstalls the installed bloat/adware. Items that redirect to the Windows Store are PIN Advertisements, they are not actually installed. Be sure to add to this list as we fine more common bloatware. To obtain the names of the installed content, use an elevated PS cmd get-appxpackage and just add a line in the arguments section in same format. Please add to this repository or edit for any recommendations, additions or adjustments.